Renamed resources to nodes and removed the map
