+++ src/components/world/node.ts
